# docs/openapi.yaml
openapi: 3.0.0
info:
  title: AgriPro API
  description: API for agricultural supply agencies management.
  version: 1.0.0
servers:
  - url: https://api.agripro.com/v1
    description: AgriPro server
...
paths:
...
  /settings:
    get:
      tags:
        - Profile & Settings
      summary: Get profile and settings
      description: Get all information of user profile and settings
      operationId: GetProfileAndSettings
      responses:
        "200":
          description: Profile and settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  userProfile:
                    $ref: "#/components/schemas/UserProfile"
                  settings:
                    type: array
                    items:
                      type: object
                      properties:
                        category:
                          $ref: "#/components/schemas/ConfigCategory"
                        key:
                          type: string
                          description: Name of attribute
                          example: "Font size"
                        value:
                          type: string
                          description: Value of attribute
                          example: "13"
        "401":
          description: Unauthorized request due to missing or invalid token.
        "500":
          description: Internal server error.
    patch:
      tags:
        - Profile & Settings
      summary: Update profile and configuration
      description: Update information of user profile and/or configuration figures.
      operationId: UpdateProfileAndSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userProfile:
                  $ref: "#/components/schemas/UserProfile"
                settings:
                  type: array
                  items:
                    type: object
                    properties:
                      category:
                        $ref: "#/components/schemas/ConfigCategory"
                      key:
                        type: string
                        description: Name of attribute
                        example: "Font size"
                      value:
                        type: string
                        description: Value of attribute
                        example: "13"
      responses:
        "200":
          description: Profile and settings updated successfully
        "400":
          description: Bad request due to validation errors or missing fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message.
        "401":
          description: Unauthorized request due to missing or invalid token.
        "500":
          description: Internal server error.
  /settings/password:
    patch:
      tags:
        - Profile & Settings
      summary: Change password
      description: Change password for a user.
      operationId: ChangePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                  description: Old password of the user
                  example: "@01Abcde"
                newPassword:
                  type: string
                  description: New password to be updated
                  example: "@01Abcdf"
              required:
                - oldPassword
                - newPassword
      responses:
        "200":
          description: Password changed successfully
        "400":
          description: Bad request due to validation errors or missing fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message.
        "401":
          description: Unauthorized request due to missing or invalid token.
        "500":
          description: Internal server error.
components:
  schemas:
...
    ConfigCategory:
      type: string
      description: Configuration category
      enum:
        - "PRINT_FORMAT"
        - "INVENTORY_PARAMS"
...
    UserProfile:
      type: object
      properties:
        agencyName:
          type: string
          description: Agency name
          example: "Đại lý A1"
        ownerName:
          type: string
          description: Owner name
          example: "Nguyễn Văn A"
        address:
          type: string
          description: Address of the agency
          example: "Mỹ Tho, Tiền Giang"
        taxCode:
          type: string
          format: number
          description: Owner's tax code
          example: "0123456789"
        phoneNumber:
          type: string
          format: number
          description: Owner's phone number
          example: "0123456789"
        email:
          type: string
          format: email
          description: Agency email / Owner email
          example: "abc@gmail.com"
        profilePicturePath:
          type: string
          description: Path to the profile picture
          example: "/profile/AG0001"